+++ src/utils/DateUtils.ts
@@ -39,7 +39,8 @@ export class DateUtils {
             year: '2-digit',
             hour: '2-digit',
             minute: '2-digit',
-            hour12: true
+            hour12: false,
+            hourCycle: 'h23'
         });
     }
 }